+++ src/index.d.ts
@@ -1,13 +1,15 @@
 import * as consts from './consts';
 import * as jadepool from './jadepool';
 import * as utils from './utils';
-import NBError = require('./NBError');
-import Task = require('./services/task');
+import * as action from './support/action';
+import NBError = require('./support/NBError');
+import Task = require('./support/task');
 
 export {
   consts,
   jadepool,
   utils,
+  action,
   NBError,
   Task
 }