+++ src/services/socketio.service.js
@@ -3,7 +3,7 @@ const socketio = require('socket.io')
 const BaseService = require('./core')
 const jp = require('../jadepool')
 const consts = require('../consts')
-const NBError = require('../NBError')
+const NBError = require('../support/NBError')
 const cryptoUtils = require('../utils/crypto')
 const logger = require('@jadepool/logger').of('Service', 'SocketIO')