+++ src/support/action/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  Context: require('./context'),
+  Action: require('./action'),
+  ActionSeries: require('./actionSeries')
+}